move command package to internal/command
The code in the command package should not be imported by other apps.
It changes often and is not intended for public use
