+++ cmd/baur/main.go
@@ -1,8 +1,6 @@
 package main
 
-import (
-	"github.com/simplesurance/baur/command"
-)
+import "github.com/simplesurance/baur/internal/command"
 
 func main() {
 	command.Execute()
