+++ internal/command/init_bashcomp.go
