+++ internal/command/ls_outputs.go
