+++ internal/command/show.go
