+++ internal/command/verify.go
